requirements.txt
openai-whisper==20240930

celery==5.4.0
# https://github.com/celery/celery/blob/main/requirements/extras/redis.txt
redis>=4.5.2,<6.0.0,!=4.5.5
# https://github.com/celery/celery/blob/main/requirements/extras/sqlalchemy.txt
SQLAlchemy>=1.4.48,<2.1
# https://github.com/sqlalchemy/sqlalchemy/blob/rel_2_0_35/setup.cfg
mysqlclient>=1.4.0
cos-python-sdk-v5==1.9.31